Refactored classes by separating validation and execution logic into distinct classes. Updated corresponding tests to improve code clarity and maintainability.
