--- .env.example
@@ -1,9 +1,0 @@
-SERVER_PORT=9090
-
-GITHUB_USERNAME=example
-GITHUB_TOKEN=example
-
-AUTH0_AUDIENCE=***
-AUTH_CLIENT_ID=***
-AUTH_CLIENT_SECRET=***
-AUTH_SERVER_URI=***
