+++ src/main/kotlin/edu/ingsis/printscriptService/services/ExecutionService.kt
@@ -1,7 +1,8 @@
-package edu.ingsis.printscriptService
+package edu.ingsis.printscriptService.services
 
 import edu.ingsis.printscriptService.DTO.ExecuteResultDTO
-import edu.ingsis.printscriptService.DTO.ResultDTO
+import edu.ingsis.printscriptService.PrintCollector
+import edu.ingsis.printscriptService.QueueInputProvider
 import edu.ingsis.printscriptService.errorHandler.ValidateErrorHandler
 import edu.ingsis.printscriptService.external.manager.ManagerAPI
 import org.springframework.beans.factory.annotation.Autowired
@@ -11,26 +12,7 @@ import java.io.ByteArrayInputStream
 import java.util.LinkedList
 
 @Component
-class Service @Autowired constructor(private val snippetManager: ManagerAPI) {
-
-    fun validate(snippet: String, version: String): ResultDTO {
-        val runner = Runner()
-        val errorHandler = ValidateErrorHandler()
-
-        runner.runValidate(
-            ByteArrayInputStream(snippet.toByteArray()),
-            version,
-            errorHandler
-        )
-
-        return ResultDTO(errorHandler.getErrors().isEmpty(), errorHandler.getErrors())
-    }
-
-    fun validateById(id: Long): ResultDTO {
-        val snippet = snippetManager.get(id).block()
-        return validate(snippet!!.snippet, snippet.version)
-    }
-
+class ExecutionService @Autowired constructor(private val snippetManager: ManagerAPI) {
     fun execute(snippet: String, version: String, input: List<String>): ExecuteResultDTO {
         val runner = Runner()
         val errorHandler = ValidateErrorHandler()
