+++ src/test/kotlin/edu/ingsis/printscriptService/ExecuteConnectWManagerE2ET.kt
@@ -1,9 +1,8 @@
 package edu.ingsis.printscriptService
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import edu.ingsis.printscriptService.DTO.ErrorDTO
 import edu.ingsis.printscriptService.DTO.ExecuteResultDTO
-import edu.ingsis.printscriptService.DTO.ResultDTO
+import edu.ingsis.printscriptService.services.ExecutionService
 import org.junit.jupiter.api.TestInstance
 import org.mockito.Mockito
 import org.springframework.beans.factory.annotation.Autowired
@@ -21,41 +20,13 @@ import kotlin.test.Test
 @AutoConfigureMockMvc
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @ActiveProfiles("test")
-class ConnectWManagerE2ET {
+class ExecuteConnectWManagerE2ET {
 
     @Autowired
     private lateinit var mockMvc: MockMvc
 
     @MockBean
-    private lateinit var service: Service
-
-    @Test
-    fun `test validate by id invalid`() {
-        val snippetId = 1L
-        val resultDTO = ResultDTO(false, listOf(ErrorDTO("Snippet is not valid")))
-
-        Mockito.`when`(service.validateById(snippetId)).thenReturn(resultDTO)
-
-        mockMvc.perform(
-            MockMvcRequestBuilders.post("/v1/validate/$snippetId")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(ObjectMapper().writeValueAsString(snippetId))
-        ).andExpect(MockMvcResultMatchers.jsonPath("$.ok").value(false))
-    }
-
-    @Test
-    fun `test validate by id valid`() {
-        val snippetId = 3L
-        val resultDTO = ResultDTO(true, emptyList())
-
-        Mockito.`when`(service.validateById(snippetId)).thenReturn(resultDTO)
-
-        mockMvc.perform(
-            MockMvcRequestBuilders.post("/v1/validate/$snippetId")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(ObjectMapper().writeValueAsString(snippetId))
-        ).andExpect(MockMvcResultMatchers.jsonPath("$.ok").value(true))
-    }
+    private lateinit var service: ExecutionService
 
     @Test
     fun `test execute by id`() {
